Don't swallow gtest failures when running tests in LLDB (#4552)
In CI, we run our various tests in LLDB as a debugger. LLDB when run
in batch mode like we are doing, always returns 0 as the exit code. It
doesn't return the exit code from the program being run. The LLDB documentation
states this is because more than 1 process can be being debugged so the
concept that GDB has for returning the error code is more problematic with
LLDB.

The result of this is that non-exception errors in CI are not being flagged
because LLDB is eating the exit code from gtest when a test fails. gest offers
the `--gtest_throw_on_failure` option that turns test failures into exceptions.

Those exceptions will be caught when we are using LLDB and reported correctly
in CI.

Long term, we have discussed adding an LLDB module with python to get us the
"return exit code". There is discussion on Stack Overflow of this approach at
https://stackoverflow.com/questions/56456724/quit-lldb-with-exit-code-of-process/76741573#76741573