Update supported Alpine to 3.20 (#4562)
This commit updates our CI to test using Alpine 3.20 for musl testing.
It also updates our release and nightly alpine ponyc images to be based
on Alpine 3.20.

Due to a change in the version of musl in Alpine 3.20, this required
a small patch LLVM 15 to compile. When we move to LLVM 16 or above,
the patch shouldn't be needed.

The new version of musl doesn't support LSF64 methods like lseek64.
In order to keep the standard library working with the new musl,
this commit adds a couple new "pony_os" functions for encapsulating
lseek and ftruncate behavior that varies by OS and on Linux, between
musl and Glibc.