Add support for pinning actors to a dedicated scheduler thread (#4547)
* Add support for pinning actors to a dedicated scheduler thread

The overall design goal and approach was to make it possible to
have pinned actors while minimizing impact to pre-existing
non-pinned actor workloads. This meant there could be no impact
on message sends (i.e. can't check to see if the receiving actor
is a pinned actor or not to decide what to do with it if it is
unscheduled). These goals were chosen because it is expected
that `pinned` actors will be a niche/small part of any pony
application's overall workload.

The approach taken has negligible performance impact to existing
scheduler logic. It adds a couple of extra checks to see if an
actor that is ready to run is a pinned actor or not and if not,
there is no other overhead involved. The scheduler quiescence
logic has an extra check for an atomic counter of pinned actors
but that is also negligible if no pinned actors are ever used.

The overall logic for pinned actors works as follows:

* The `main` thread is dedicated to running pinned actors (and
only pinned actors). This thread previously initialized the
runtime and then sat around waiting for all schedulers to reach
quiescence so now it runs pinned actors in the meantime if there
are any.
* The `pinned actor thread` (`main`) runs a custom run loop for
pinned actors that does not participate in work stealing or any
other normal scheduler messaging except for unmuting messages and
the termination message. It also will only ever run `pinned` actors
and any non-`pinned` actors will get pushed onto the `inject` queue.
* Normal schedulers will only ever run non-`pinned` actors and any
`pinned` actors will get pushed onto the `pinned actor thread`'s
queue.
* From an api perspective, there is now an `actor_pinning` package
in the stdlib. An actor can request to be pinned, check that it
has successfully been pinned (so that it can safely do whatever it
needs to do while pinned), and request to be unpinned.

While the above is not necessarily the most efficient way to run
`pinned` actors, it meets the original design goals of making it
possible while minimizing impact of pre-existing non-pinned actor
workloads.

* Fix unused variable error

* Fix dtrace undeclared variable error

* Fix windows pinned-actor test linking error

* Add greedy actor caveat to actor pinning package

* Add release notes

* Rename `pin` to `request_pin` and `unpin` to `request_unpin`

and also update the package documentation to clarify that `pinning`
is not an immediate action

* update release notes

* Use correct function names now that they've been renamed

* update caveat

Co-authored-by: Sean T Allen <sean@seantallen.com>

* Fatten up the release notes

* make pinned actor thread participate in CNF/ACK for termination

* add option to pin pinned actor thread

---------

Co-authored-by: Sean T Allen <sean@seantallen.com>