Deterministic cerealisation for cross binary communication
Prior to this commit, serialisation used the `type_id` of the
types that it serialised and the `type_id`s assigned to types
can vary depending on the program being compiled. This makes it
impossible for cross binary serialisation even if the same pony
version and types are used.

This commit changes things so that serialisation now uses a
determnistically assigned `serialise_id` instead of the `type_id`.
This new `serialise_id` is guaranteed to be the same for the same
pony version and type (see details for caveats).

The changes include:

* adding a `size_t` size `serialise_id` to `pony_type_t`
* hashes the reach_type_t->name to generate either a 63 bit or
  31 bit `serialise_id` depending on target platform
* the hash for the `serialise_id` uses a `key` that is based on the
  blake2 encoding of the pony version, the target data model, and
  the endianness (i.e. `pony-0.58.7-ac1c51b69-lp64-le`). This is
  to ensure that only programs with the same `key` will have the
  same `serialise_id` to prevent silent data corruption during
  deserialisation due to subtle differences in platforms. This
  means that cross binary serialisation will only work for binaries
  of the same bit width (32 bit vs 64 bit), data model (ilp32, lp64,
  or llp64), and endianness (big endian or little endian)
* serialisation now uses the `serialise_id` instead of the `type_id`
* deserialisation now uses the `serialise_id` to look up the
  `type_id` to get the entry in the descriptor table instead of
  looking up the entry in the descriptor table directly based on
  `type_id`