Fix compiler crash from `match` with extra parens around `let` in tuple (#4595)
The compiler would crash if an extra set of parens were added to a
single element tuple match of a union type. For example,
if `type Foo is (I32 | (I32, I32))` is being matched, then a match
clause like `| (123, (let x: I32)) => x` would crash the compiler.
This was caused by the codegen for the match not handling the extra
SEQ node in the AST created by the parens. This change now skips
any extra SEQ nodes when performing codegen for tuple matches.

Fixes #4412