resurrect pony_always_assert and decouple it from pony_ndebug
pony_always_assert was added at some point to help always get
backtraces from assertions from ponyc (even in release builds) but
it was disabled because it was including debug code in release
builds unnecessarily.

this commit ensures that pony_always_assert now only impacts
pony_assert (and only for libponyc not libponyrt) to make sure
that debug code will not be included in release builds and also
only enables pony_always_assert for libponyc and libponyc.tests
(and not libponyrt).