Fix cycle detector rescheduling bug introduced in #4607
PR #4607 introduced a bug causing the cycle detector to be
rescheduled when it didn't need to be rescheduled. this sometimes
resulted in the cycle detector being run concurrently in two
different scheduler threads which would cause a segfault or
assertion failure.

This commit fixes the logic issue so that the cycle detector will
only be recscheduled if it indicated that it needed to be
rescheduled to ensure that only a single copy of the cycle detector
will be running at a time as expected.