Don't run GC for zombie actors when they are blocked (#4617)
Prior to this commit, if the cycle detector was enabled all actors
would run a GC cycle before sending a block message to the cycle
detector. This is helpful for actors that are waiting for additional
work to be able to release some memory that they don't need but
wasteful for zombie actors that the cycle detector will destroy
as soon as it receives the block message for them.

This commit changes the logic so that the GC cycle no longer run
for zombie actors when they block because the cycle detector will
destroy the actor soon anyway.