Remove invalid asserts from ASIO system
There's an ABA condition that can exist with how the ASIO system
works in the runtime. There's guard
conditions in various functions to ditch early for scenarios where
this is encountered.

A few years ago, some asserts were added to those guard clauses.
Unfortunately, that defeats the purpose of the guards and will
eventually result in assertions being triggered under load with a
debug runtime.

Without completely rewriting how the ASIO system works to be ABA
friendly, these asserts are a bad idea. I don't feel up to doing a
rewrite at this time, so I am removing the asserts.