Update the QEMU used in our cross runner (#4631)
We are getting some weird failures in our cross CI so this is an
easy thing to attempt to fix before we call it a day and give up
on the 32-bit testing.

The lastest in the 7 series had a bus error so I am going with this, the latest in the 6 series.