Add live asio event tracking for debugging (#4644)
This might not live on in this fashion for a long time. This at
the moment exists for the pony_asserts. However, to make the
cycle detector work better, we do want to know if an actor has
any live asio events. If it does, we should never mark it as
blocked as it isn't. A live asio event means that the actor is
capable of getting messages and should never be evaluated for
being in a cycle. That change however, is for another day.