Fix ASIO thread scheduler index (#4646)
Prior to this commit, the ASIO thread index was hardcoded via
constant in multiple places. Sometimes this was correct but other
times it wasn't (because the function was being called by an actor
instead of the ASIO theead).

This commit changes things so that the ASIO thread index can be
retrieved the same way as any other scheduler thread index and
replace all the hardcoded constant references with a call to the
pony_scheduler_index() function so that the scheduler index is
always the correct one.