Make sure systematic testing doesn't switch to suspended thread (#4647)
* Make sure systematic testing doesn't switch to suspended thread

Prior to this commit, the systematic testing next thread logic
would consider switching to the pinned actor thread even if it
was suspended. This was very wasteful since the suspended pinned
actor thread would not do any meaningful work and almost immediately
re-suspend itself.

This commit changes things so that the next thread logic properly
takes into account whether the pinned actor thread is suspended
or not when selecting the next thread to activate.

* add release notes