Add ability to trace runtime events in chromium json format
Runtime tracing is helpful in understanding how the runtime
works and for debugging issues with the runtime.

The implementation hopefully is flexible and extensible in order
to be able to add more events for tracing as needed.

The tracing can either write to file in the background or work in
a "flight recorder" mode where events are stored into in memory
circular buffers and written to stderr in case of abnormal program
behavior (SIGILL, SIGSEGV, SIGBUS, etc).

Chromium json format is documented at:

https://docs.google.com/document/d/1CvAClvFfyA5R-PhYUmn5OOQtYMH4h6I0nSsKchNAySU/preview?tab=t.0

and trace files can be viewed with the perfetto trace viewer:

https://perfetto.dev/

inspired by golang's tracing and verona's flight recorder.