Split runtime stress tests apart into different workflows
This will allow us now that we have a "run manually option" to
more easily target a subset of tests. Additionally, the GitHub UI
for the stress tests was really hard to navigate with the old set
up. This will allow for a more focused break down in the UI.