Fix windows ASIO clobbering scheduler index bug introduced in #4646 (#4652)
In #4646, we started using a dedicated scheduler index for the ASIO
thread. This was generally a good thing but on windows due to how
IOCP works, we could accidentally overwrite the scheduler index for
a normal scheduler if it happened to be the one receiving the ASIO
callback.

This commit fixes the bug to ensure that we will not overwrite the
scheduler index during ASIO callbacks on windows.