Fix user-after-free in actor run stop tracing events
Prior to this commit, the actor run stop tracing events could
potentially result in a use-after-free if the actor had already
been destroyed (by the cycle detector or the fast actor reaping
logic) by the time the actor run stop tracing event was issued.

This commit changes the logic so that the actor run start and
actor run stop tracing events are now issued from within
`ponyint_actor_run` and explicitly issued at a time before the
actor is flaged to be destroyed to ensure that there is no
possibility of a use-after-free error.