Use atomics for tracing flight recorder events
Prior to this commit, the flight recorder buffer events were stored
without the use of atomics. While this is not an issue for strong
consistency systems like x86, it can potentially lead to the
flight recorder missing the most recent events that might have
occurred in schedulers on weak consistency systems like arm.

This commit changes the logic for flight recorder events to always
use atomics when storing and retrieving them to try to ensure that
the flight recorder events are as up to date and accurate as
possible if/when a program error occurs.