Don't force tracing flight recorder output to stderr anymore
Prior to this commit, the flight recorder events would always be
dumped to `stderr`. While this was useful, sometimes the number
of events could overflow the scrollback buffer in a users' terminal.
Additional, It is very likely that most users would want to take
the events from the flight recorder and process them further via
tools like `jq` or visualize them in tools like `perfetto`, both
of which would require the user to copy/paste the events into a
file.

This commit makesit so the flight recorder dumps the events out
to whatever the `--ponytracingoutput` cli option specifies which
the user can still control to be `stderr` if desired.