Add partial support for Windows on Arm64
This adds checking on PR for Windows on Arm64. I haven't been
able to get LLDB working correctly for tests so that is off.

This also adds building nightlies so we can see if there are issues
in that workflow.

No changelog entry yet as this is an exploratory change with others
to follow if this works out.