Make sure LLVM patch directory changes are caught despite caching (#4698)
When running `make libs`, we check a hash of the files in the `lib/llvm/patches` directory to make sure we are building in a known state.  However, since we cache the LLVM build in CI, this check is not necessarily run when the files in the `patches` directory change.

This change updates the desired hash of the patch files, and updates LLVM caching to include the patches directory in its key.