Set explicit linker when testing 32 bit cross compiled architectures
For reasons that are unknown to me, the ld.bfd installed in our 32-bit
testing environments does not skip over incompatible libraries when attempting
to link.

This means if bfd is used, that we can't link programs because it first
finds the host environment pony runtime and fails. The ld.gold that is in
the environment correctly skips over and finds the correct one.

I spent more time than I should have on trying to fix so that it doesn't
ever find the host pony runtime but never got that working. This is the
"make it work fix".

This fix is needed because gold has been ejected from binutils as Google
has abandoned work on it and no one else has come forward to support so
GNU is back to bfd as "the GNU linker".

As the lack of gold rolls out to various Linux distros, we are going to
need to stop having gold be the default Linux linker. By explicitly setting
gold in these two environments, they will continue to work when that change
happens.