Use more appropriate arch when doing arm64 builds
Our armv8 settings came because that it what was reported as working
on m1 macs. However, I had Joe test and armv8-a works fine. -a will
be targeted to a more specific set of arm processors including aarch64.

There are "smaller" arm64 processors that are included with armv8 that
no one is likely to try to use our prebuilts on so armv8-a is a better
floor for us to set.