Add Multiplatform Nightly musl Images
This introduces a number of changes for our container image building.

It introduces the eventual deprecation of `latest` as a tag. At the moment,
our use of `latest` is misleading. That is supposed to be "most recent" but
we don't use it as such. For us it is "most recently nightly".

This commit will be the start of us having "nightly" images that are explicitly
a nightly build.

In addition the tag name changes, this also is our first multiplatform image. The
build process for this is as simple as I could make it. We have a new workflow for
building the nightly image. It is set to have a concurrency of 1 so, as events arrive
from Cloudsmith, we will build a new version of the image with whatever is the latest
nightly arm64 and latest nightly am64 version of ponyc that is available in Cloudsmith.

This means that as we upload to Cloudsmith each night we will create the image twice.
If the arm64 synchronized message arrives first and then the amd64 one we will have the
following happen:

- arm64 message arrives
- Build a new "nightly image" with today's arm64 and yesterday's amd64
- amd64 message arrives
- Build a new "nightly image" with today's arm64 and today's arm64

The same pattern would be applied to releases.

At the time the message arrives, the image will be built using the most recent version
of ponyc for each platform that is available. This has a couple nice qualities to it:

- If one failed to build, we still make the one that succeeded available.
- It makes our configuration relatively simple

Without this "cross over", we would need to wait for each message to have arrived and then
build the final image. That is some tricky yaml that we should want to avoid doing if
possible.