Cause registry_package events to get sent
It appears that if we push an image with the default token, that
registry_package events are not sent. This is like how you can't
trigger a new workflow from an action using the default github
token like using that token to push to main.

This commit updates all our ponyc images that would require multiplatform
builds aka all the Linux ones to use a PAT from ponylang-main to push
thereby allowing registry_package events to occur.