Stop building glibc based ponyc images
Prior to this change, every night and for every release, we built
2 sets of ponyc container images. The first was musl based. The second
was glibc based images.

The glibc based images were based on an ever changing version of
Ubuntu (most recent LTS release). These images had minimal utility.

They could only be used to link for Linux distros that had the
same libraries installed as the build environment. Whereas the musl
images could be used to statically link using musl and create binaries
that could be used on any Linux.

Given the lack of utility for a wide variety of people, we have stopped
building the glibc based images and will only have a single nightly and
release image going forward. Those images will be using musl via Alpine
Linux.

If you were using the glibc based images, we can point you to the Dockerfiles
we were using as they are quite straightfoward and should be easy for
anyone to incorporate the creation of into their own CI process.