Stop the "registry package event" test
I had set up a workflow to dump information from registry package
events in case we wanted to use them as part of our multiplatform
image work.

In the end, we didn't use those events, so I have turned off the test
script. Additionally, in order to get the events to fire, we needed
to be using a token for pushing to GHCR that wasn't `secrets.GITHUB_TOKEN`.
I have reverted back to GITHUB_TOKEN as we don't need the "trigger new
events" that having a customized PAT enabled.