Set vendored LLVM triples to clang -dumpmachine
The work we've been doing to integrate lld has exposed that the
triple that is being returned by our vendored llvm was the generic
default, even when it should be something else.

This hasn't come up before, as we've not tested for things like
target_is_musl() until this point.

As we require a "system clang" to compile llvm, we can use this
to set the default value of LLVM_HOST_TRIPLE and
LLVM_DEFAULT_TARGET_TRIPLE.